+++ apps/bot/src/domains/player/service.ts
@@ -1,7 +1,9 @@
-import { db, type Player } from '@one-piece/db';
+import { db, type DbOrTransaction, type Player, type Zone } from '@one-piece/db';
 
 import { sanitizeName } from '../../shared/sanitize-name.js';
 import * as characterRepository from '../character/repository.js';
+import { bucketIdFromTimestamp } from '../event/engine/bucket.js';
+import * as historyRepository from '../history/index.js';
 import { findOrCreateShip } from '../ship/service.js';
 
 import { assertNameNotEmpty, assertNameWithinMaxLength } from './guards/index.js';
@@ -38,4 +40,39 @@ export async function renamePlayer(playerId: number, rawName: string): Promise<P
     await characterRepository.updatePlayerAsCharacterNickname(playerId, sanitizedName, transaction);
     return updated;
   });
+}
+
+type RecordZoneChangeInput = {
+  playerId: number;
+  newZone: Zone;
+  at?: Date;
+  client?: DbOrTransaction;
+};
+
+async function recordZoneChangeWithClient(input: RecordZoneChangeInput, client: DbOrTransaction): Promise<void> {
+  const at = input.at ?? new Date();
+  const bucketId = bucketIdFromTimestamp(at);
+  const currentPlayer = await playerRepository.findByIdOrThrow(input.playerId, client);
+  const from = currentPlayer.currentZone;
+
+  if (from === input.newZone) return;
+
+  await playerRepository.updateZone(input.playerId, input.newZone, client);
+  await historyRepository.appendHistory({
+    type: 'player.zone_changed',
+    actorPlayerId: input.playerId,
+    bucketId,
+    payload: { from, to: input.newZone },
+    client,
+    occurredAt: at,
+  });
+}
+
+export async function recordZoneChange(input: RecordZoneChangeInput): Promise<void> {
+  if (input.client) {
+    await recordZoneChangeWithClient(input, input.client);
+    return;
+  }
+
+  await db.transaction(async (transaction) => recordZoneChangeWithClient(input, transaction));
 }
