+++ packages/db/src/index.ts
@@ -1,4 +1,4 @@
-export { closeDb, db, type Db, type DbOrTransaction, type DrizzleTx } from './client.js';
+export { closeDb, db, type Db, type DbOrTransaction, type Transaction } from './client.js';
 export { MAX_CHARACTER_NAME_LENGTH } from './shared/constants.js';
 export { type JSONFromSQL } from './shared/types.js';
 export * from './schema.js';
